lppublish: new package, factored from cmd/charmload
